requirements.txt
contourpy==1.1.1
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.42.1
importlib-resources==6.0.1
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.1
pyDOE2==1.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
tzdata==2023.3
zipp==3.17.0
